chore: separado Span de Metadata, além de algumas limpezinhas
