+++ program/includes/Core/Utils/Metadata.hpp
@@ -1,10 +1,8 @@
 #pragma once
 
-#include <cstddef> // size_t
-#include <iterator>
-#include <array>
+#include "Core/Utils/Span.hpp"
 
-#define ENUM_VALUE(name) { #name, static_cast<int>(T::name) }
+#include <cstring>
 
 namespace core
 {
@@ -23,63 +21,19 @@ namespace core
 		Delegate
 	};
 
-	/**
-	 * @brief Classe básica minimalista para Span.
-	 **/
-	template<typename T>
-	class Span {
-	public:
-	    using element_type    = T;
-	    using value_type      = typename std::remove_cv<T>::type;
-	    using size_type       = std::size_t;
-	    using pointer         = T*;
-	    using reference       = T&;
-	    using iterator        = T*;
-	    using const_iterator  = const T*;
-
-	    constexpr Span() noexcept : data_(nullptr), size_(0) {}
-	    constexpr Span(T* ptr, size_type count) noexcept : data_(ptr), size_(count) {}
-	    constexpr Span(T* first, T* last) noexcept : data_(first), size_(last - first) {}
-
-	    template<size_t N>
-		constexpr Span(std::array<value_type, N>& arr) noexcept
-		    : data_(arr.data()), size_(N) {}
-
-		// Construtor a partir de std::array const
-		template<size_t N>
-		constexpr Span(const std::array<value_type, N>& arr) noexcept
-		    : data_(arr.data()), size_(N) {}
-
-	    template<size_t N>
-	    constexpr Span(T (&arr)[N]) noexcept : data_(arr), size_(N) {}
-
-	    constexpr iterator begin() const noexcept { return data_; }
-	    constexpr iterator end()   const noexcept { return data_ + size_; }
-
-	    constexpr reference operator[](size_type idx) const { return data_[idx]; }
-	    constexpr size_type size() const noexcept { return size_; }
-	    constexpr bool empty() const noexcept { return size_ == 0; }
-	    constexpr pointer data() const noexcept { return data_; }
-
-	private:
-	    T* data_;
-	    size_type size_;
-	};
-
-
 	struct EnumEntry 
 	{
-	    const char* name;
-	    int value;
+	    const char* Name;
+	    int Value;
 	};
 
 	struct TypeDescriptor
 	{
-	    MetaType type;
-	    const char* typeName;
+	    MetaType Type;
+	    const char* TypeName;
 
 	    // Para enums
-	    Span<const EnumEntry> enumEntries;
+	    Span<const EnumEntry> EnumEntries;
 	};
 
 	/**
@@ -89,13 +43,46 @@ namespace core
 	struct EnumRegistry; // default vazio
 
 	template<typename T>
-	inline constexpr TypeDescriptor MakeEnumDescriptor(const char* typeName) 
+	inline constexpr TypeDescriptor MakeEnumDescriptor(const char* typename_) 
 	{
 	    return TypeDescriptor
 	    {
 	        MetaType::Enum,
-	        typeName,
+	        typename_,
 	        EnumRegistry<T>::entries
 	    };
+	}
+
+	/**
+	 * @brief Converte um enumerador para string. 
+	 **/
+	inline const char* EnumToString(const TypeDescriptor& descriptor, int value)
+	{
+		if (descriptor.Type != MetaType::Enum) return nullptr;
+		for (auto& e : descriptor.EnumEntries)
+		{
+			if (e.Value == value)
+			{
+				return e.Name;
+			}
+		}
+		return nullptr;
+	}
+
+	/**
+	 * @brief Converte uma string para enumerador. 
+	 **/
+	inline bool StringToEnum(const TypeDescriptor& descriptor, const char* str, int& value)
+	{
+		if (descriptor.Type != MetaType::Enum) return false;
+		for (auto& e : descriptor.EnumEntries)
+		{
+			if (strcmp(str, e.Name) == 0)
+			{
+				value = e.Value;
+				return true;
+			}
+		}	
+		return false;
 	}
 }
